- Replaced "sizeX" with "width" and "sizeY" with "height" - Use imageQuality_t instead of imageQuality - Corrected bunch of implicit integer conversions - All the API constructors involving an implementation class now are private - Added virtual destructor to all API classes - Separated API Memory class in ReadMemory and ReadWriteMemory -

+++ library/implementation/fileStreamImpl.cpp
@@ -240,7 +240,7 @@ void fileStreamOutput::write(size_t startPosition, const std::uint8_t* pBuffer, 
 
     std::lock_guard<std::mutex> lock(m_mutex);
 
-	::fseek(m_openFile, startPosition, SEEK_SET);
+    ::fseek(m_openFile, (long)startPosition, SEEK_SET);
 	if(ferror(m_openFile) != 0)
 	{
         IMEBRA_THROW(StreamWriteError, "stream::seek failure");
@@ -270,7 +270,7 @@ size_t fileStreamInput::read(size_t startPosition, std::uint8_t* pBuffer, size_t
 
     std::lock_guard<std::mutex> lock(m_mutex);
 
-	::fseek(m_openFile, startPosition, SEEK_SET);
+    ::fseek(m_openFile, (long)startPosition, SEEK_SET);
 	if(ferror(m_openFile) != 0)
 	{
 		return 0;
